Server/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
transformers==4.35.0
torch==2.0.1
faiss-cpu==1.7.4
google-cloud-storage==2.10.0
pyarrow==13.0.0
numpy==1.23.5
openai==0.27.0
google-cloud-secret-manager==2.16.1
llama-index==0.5.1  # Assuming llama-index is installed via pip
pinecone-client==2.2.0  # Assuming this is the package name for pinecone
